+++ pkg/innerring/processors/container/processor.go
@@ -179,6 +179,16 @@ func (cp *Processor) ListenerNotaryParsers() []event.NotaryParserInfo {
 	p.SetRequestType(fschaincontracts.AddContainerStructsMethod)
 	p.SetParser(containerEvent.RestoreAddStructsRequest)
 
+	// set attribute
+	p.SetRequestType(fschaincontracts.SetContainerAttributeMethod)
+	p.SetParser(containerEvent.RestoreSetAttributeRequest)
+	pp = append(pp, p)
+
+	// remove attribute
+	p.SetRequestType(fschaincontracts.RemoveContainerAttributeMethod)
+	p.SetParser(containerEvent.RestoreRemoveAttributeRequest)
+	pp = append(pp, p)
+
 	return pp
 }
 
@@ -254,6 +264,16 @@ func (cp *Processor) ListenerNotaryHandlers() []event.NotaryHandlerInfo {
 
 		cp.log.Info("notary tx migrating containers' protobuf->struct signed successfully")
 	})
+
+	// set attribute
+	h.SetRequestType(fschaincontracts.SetContainerAttributeMethod)
+	h.SetHandler(cp.handleSetAttribute)
+	hh = append(hh, h)
+
+	// remove attribute
+	h.SetRequestType(fschaincontracts.RemoveContainerAttributeMethod)
+	h.SetHandler(cp.handleRemoveAttribute)
+	hh = append(hh, h)
 
 	return hh
 }